# PERF: speed up default GeoDataFrame constructor from internal manager