Internally in pandas methods, often `_constructor` (and thus `GeoDataFrame(..)` is called with already a fully formed BlockManager object without any other arguments (basically just to wrap this data in the class / subclass). In those cases, our `GeoDataFrame` has quite some overhead.

For example, making a copy of a GeoDataFrame:

```
In [1]: df = geopandas.read_file(geopandas.datasets.get_path("naturalearth_lowres"))

In [2]: %timeit df.copy()
681 µs ± 6.06 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)`  # <- main
98.4 µs ± 1.35 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)  # <- PR
```

This is about relatively small timings in general, but because this can be called multiple times in an operation, this can actually becomes noticeable (eg in `sjoin` on larger data, where the sjoin was taking 1-2 seconds, this was visible significantly in the profile).